+++ refinery-core/src/test/java/org/wikimedia/analytics/refinery/core/TestLegacyPageviewDefinition.java
@@ -33,6 +33,7 @@ public class TestLegacyPageviewDefinition {
     )
     public void testIsLegacyPageview(
         String test_description,
+        String project,
         boolean is_pageview,
         boolean is_legacy_pageview,
         boolean is_app_pageview,